Abstract UnitControl to block-editor component
The block-editor variant of UnitControl has add_theme_support built in
